starknet_committer,starknet_os: relocate commitment infos to starknet_committer
